requirements.txt
asgiref==3.5.2
Django==4.1.2
django-debug-toolbar==3.7.0
djangorestframework==3.14.0
drf-nested-routers==0.93.4
mysqlclient==2.1.1
python-decouple==3.6
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
